+++ src/app/auth/signin/signin.component.ts
@@ -40,7 +40,9 @@ export class SigninComponent implements OnInit, OnDestroy {
   }
 
   signup() {
-    this.service.signup(this.signupForm.get('email').value, this.signupForm.get('password').value);
+    this.service.signup(this.signupForm.get('email').value, this.signupForm.get('password').value).then(() => {
+      this.service.signin(this.signupForm.get('email').value, this.signupForm.get('password').value);
+    });
   }
 
   signout() {